[Add] basic Input bindings (Analog, Digital and Debug key)
